update all tables for new standings routing
