+++ app/api/matches/route.ts
@@ -1,10 +1,16 @@
 import { NextResponse } from 'next/server';
-import { Matches } from '@/interfaces/players';
+import { Match } from '@/interfaces/match';
+import standing from '@/data/league-details.json';
 
-async function fetchData(): Promise<{ matches: Matches[] }> {
+async function fetchData(): Promise<{ matches: Match[] }> {
   try {
-    const res = await fetch('https://draft.premierleague.com/api/league/5525/details');
-    return await res.json();
+    if (process.env.NODE_ENV === 'development') {
+      const res = standing;
+      return await res;
+    } else {
+      const res = await fetch('https://draft.premierleague.com/api/league/5525/details');
+      return await res.json();
+    }
   } catch (err) {
     console.error(err);
     throw err;
