+++ tailwind.config.js
@@ -11,6 +11,12 @@ module.exports = {
   ],
   theme: {
     extend: {
+      screens: {
+        sm: '640px',
+        md: '768px',
+        lg: '1024px',
+        xl: '1280px',
+      },
       colors: {
         premPurple: '#310639',
         premGreen: '#75fa95',
